src/api/requirements.txt
fastapi==0.111.0
aiohttp==3.9.5
pydantic==2.7.2
pydantic-settings==2.2.1
pyautogen==0.2.35
nest_asyncio==1.6.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-distro==0.47b0
httpx==0.27.0
asyncpg==0.29.0
pandas==2.2.2
async-lru==2.0.4
pyinstrument==4.7.2
asgi-correlation-id==0.1.1
python-json-logger==2.0.7
azure-monitor-opentelemetry==1.6.1
fastapi-profiler==1.4.1
msal==1.30.0
azure-identity==1.17.1